Update some repositories

+++ api.schedule/schedule-educational-program-service/cmd/app/internal/services/discipline/repository.go
@@ -7,7 +7,7 @@ import (
 
 type Repository interface {
 	GetAll(context.Context) (*[]model.Discipline, error)
-	GetByID(context.Context, int) (*model.Discipline, error)
+	GetByID(context.Context, int64) (*model.Discipline, error)
 	Save(context.Context, *model.Discipline) error
-	Delete(context.Context, int) error
+	Delete(context.Context, int64) error
 }
